Cast the *result* to float
I was seeing console log errors complaining that supplier_price needed to be an
integer instead of a string. Checking the props, employee_price was indeed a
number but supplier_price was in double quotes.

Looking at the transformer, it seemed like perhaps the float cast might be
getting applied to the in_array part rather than the actual values. The
fix I put into place ensured the float cast would be on the result of
the entire expression, not just the in_array part.